+++ django/core/apps.py
@@ -1,8 +1,0 @@
-from django.apps import AppConfig
-
-class MyAppConfig(AppConfig):
-    name = 'core'
-    verbose_name = "FromEdwin Core"
-
-    def ready(self):
-        import core.signals # Load signals