+++ src/core/apps.py
@@ -1,0 +1,9 @@
+from django.apps import AppConfig
+
+class MyAppConfig(AppConfig):
+    name = 'core'
+    verbose_name = "FromEdwin Core"
+
+    def ready(self):
+        import core.signals # Load signals
+        import django_celery_beat.schedulers